resume_ai_pipeline/requirements.txt
pymupdf==1.22.0
pdf2image==1.16.3
pytesseract==0.3.10
python-docx==0.8.11
Pillow>=10.3.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
nltk==3.8.1
scikit-learn==1.3.2
